+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ## Changelog
 
 * Removed hardcoded default return nodes when failing an Enemy Ambush (now uses point 0).
+* Dynamic resource lists! See `/src/world_maps/data/resource_lists.h` to change these.